# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Report System API",
        "description": "API documentation for the report system backend. Most school-scoped endpoints require an owner Sanctum token with the `permit:owner` ability.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/auth/sign-up": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a school owner",
                "description": "Creates an owner client and their school, then returns a Sanctum token.",
                "operationId": "7ac9719bf1f0864eedbdb9ea3ded591e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "schoolName",
                                    "address",
                                    "city",
                                    "type",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ama Owner",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "owner@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password1!",
                                        "minLength": 8
                                    },
                                    "schoolName": {
                                        "type": "string",
                                        "example": "Ridge SHS",
                                        "maxLength": 255
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "12 Independence Ave",
                                        "maxLength": 255
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Accra",
                                        "maxLength": 255
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "private",
                                        "enum": [
                                            "private",
                                            "public"
                                        ]
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0240000000",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registered successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AuthTokenData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "School": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Ridge SHS"
                    },
                    "address": {
                        "type": "string",
                        "example": "12 Independence Ave"
                    },
                    "city": {
                        "type": "string",
                        "example": "Accra"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "private",
                            "public"
                        ]
                    },
                    "imageUrl": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "0240000000"
                    },
                    "motto": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "ownerId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
            "Client": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Ama Owner"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "owner@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "owner",
                            "teacher",
                            "student"
                        ]
                    },
                    "schools": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/School"
                        }
                    }
                },
                "type": "object"
            },
            "AuthTokenData": {
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|abcdef..."
                    },
                    "token_type": {
                        "type": "string",
                        "example": "Bearer"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token. Send as: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}